refactor: remove destroy controller out of common
DestroyController was used in exactly one place, so it's hardly common.
Remove it and put it at the correct location.
