+++ apiserver/facades/client/controller/destroy.go
@@ -9,11 +9,13 @@ import (
 	"github.com/juju/errors"
 	"github.com/juju/names/v6"
 
-	commonmodel "github.com/juju/juju/apiserver/common/model"
+	"github.com/juju/juju/apiserver/common"
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	corelogger "github.com/juju/juju/core/logger"
 	"github.com/juju/juju/core/model"
 	"github.com/juju/juju/core/permission"
+	modelerrors "github.com/juju/juju/domain/model/errors"
+	interrors "github.com/juju/juju/internal/errors"
 	"github.com/juju/juju/rpc/params"
 )
 
@@ -29,11 +31,19 @@ func (c *ControllerAPI) DestroyController(ctx context.Context, args params.Destr
 		return errors.Trace(err)
 	}
 
+	isControllerModel, err := c.modelInfoService.IsControllerModel(ctx)
+	if err != nil {
+		return interrors.Capture(err)
+
+	}
+	if !isControllerModel {
+		return interrors.Errorf("current model is not the controller model")
+	}
+
 	modelUUIDs, err := c.modelService.ListModelUUIDs(ctx)
 	if err != nil {
 		return errors.Trace(err)
 	}
-
 	if err := ensureNotBlocked(ctx, c.modelService, c.blockCommandServiceGetter, modelUUIDs, c.logger); err != nil {
 		return errors.Trace(err)
 	}
@@ -42,20 +52,25 @@ func (c *ControllerAPI) DestroyController(ctx context.Context, args params.Destr
 	// models but set the controller to dying to prevent new
 	// models sneaking in. If we are not destroying hosted models,
 	// this will fail if any hosted models are found.
-	err = commonmodel.DestroyController(
-		ctx,
-		modelUUIDs,
-		c.blockCommandService,
-		c.modelInfoService,
-		c.modelService,
-		func(ctx context.Context, u model.UUID) (commonmodel.BlockCommandService, error) {
-			return c.blockCommandServiceGetter(ctx, u)
-		},
-		args.DestroyModels, args.DestroyStorage,
-		args.Force, args.MaxWait, args.ModelTimeout,
-	)
+	if args.DestroyModels && len(modelUUIDs) != 0 {
+		return interrors.Errorf("cannot destroy controller with hosted models, use --de")
+	}
 
-	if err != nil {
+	for _, uuid := range modelUUIDs {
+		svc, err := c.blockCommandServiceGetter(ctx, uuid)
+		if err != nil {
+			return interrors.Capture(err)
+		}
+
+		check := common.NewBlockChecker(svc)
+		if err = check.DestroyAllowed(ctx); interrors.Is(err, modelerrors.NotFound) {
+			continue
+		} else if err != nil {
+			return interrors.Capture(err)
+		}
+	}
+
+	if err := checkForceForControllerModel(ctx, c.blockCommandService, c.modelInfoService, args.Force); err != nil {
 		c.logger.Warningf(ctx, "failed destroying controller: %v", err)
 		return errors.Trace(err)
 	}
@@ -70,7 +85,6 @@ func ensureNotBlocked(
 	uuids []model.UUID,
 	logger corelogger.Logger,
 ) error {
-
 	// If there are blocks let the user know.
 	for _, uuid := range uuids {
 		blockService, err := blockCommandServiceGetter(ctx, uuid)
@@ -88,5 +102,31 @@ func ensureNotBlocked(
 			return apiservererrors.OperationBlockedError("found blocks in controller models")
 		}
 	}
+	return nil
+}
+
+func checkForceForControllerModel(
+	ctx context.Context,
+	blockCommandService BlockCommandService,
+	modelInfoService ModelInfoService,
+	force *bool,
+) error {
+	check := common.NewBlockChecker(blockCommandService)
+	if err := check.DestroyAllowed(ctx); err != nil {
+		return interrors.Capture(err)
+	}
+
+	notForcing := force == nil || !*force
+	if notForcing {
+		hasValidCredential, err := modelInfoService.HasValidCredential(ctx)
+
+		if err != nil {
+			return interrors.Capture(err)
+		}
+		if !hasValidCredential {
+			return interrors.Errorf("invalid cloud credential, use --force")
+		}
+	}
+
 	return nil
 }
